+++ NEWS.md
@@ -2,11 +2,19 @@ ChangeLog for SamsElves Package
 
 --------------------------------------------------------------------------------
 
+## 2024-10-22 v.0.3.14
+
+- improved:
+  - build_metadata_shell finds number of cores using 'nproc'
+
+
+
 ## 2024-10-22 v.0.3.13
 
 - bugfix:
   - aggregate_from_children_to_parents 
     - assert complete cases
+
 
 ## 2024-10-22 v.0.3.12